+++ application/views/profiles/edit.php
@@ -17,60 +17,45 @@
         <?php
             }
         ?>
+        <div class="alert alert-dismissible alert-secondary" style="margin-top: 16px">
+            <?= $this->lang->line('mandatory_fields'); ?>
+        </div>
         <div class="form-group">
             <!-- <div class="col-md-3"> -->
-                <label for="name" class="col-sm-4 col-form-label"><strong><?= $name; ?></strong>:
+                <label for="name" class="col-sm-4 col-form-label"><strong><?= $name; ?></strong>*:
                 <input type="text" class="form-control" id="name" name="name" value="<?= $profile_name; ?>"></label>
-                <label for="surname" class="col-sm-4 col-form-label"><strong><?= $surname; ?></strong>:
+                <label for="surname" class="col-sm-4 col-form-label"><strong><?= $surname; ?></strong>*:
                 <input type="text" class="form-control" id="surname" name="surname" value="<?= $profile_surname; ?>"></label><br>
                 <fieldset class="form-group">
                     <label class="form-label"><strong><?= $gender; ?></strong>:
-                        <?php
-                            if ($profile_gender === $this->lang->line('profile_gender_m')) {
-                        ?>
                         <div class="form-check"> 
                             <label for="gender" class="form-check-label">
-                                <input type="radio" class="form-check-input" id="gender_m" name="gender" value="m" checked>
+                                <input type="radio" class="form-check-input" id="gender_m" name="gender" value="m" <?php if ($profile_gender == $this->lang->line('profile_gender_m')) { echo 'checked'; } ?>>
                                 <?= $this->lang->line('profile_gender_m'); ?>
                             </label>
                         </div>
                         <div class="form-check"> 
                             <label for="gender" class="form-check-label">
-                                <input type="radio" class="form-check-input" id="gender_f" name="gender" value="f">
+                                <input type="radio" class="form-check-input" id="gender_f" name="gender" value="f" <?php if ($profile_gender == $this->lang->line('profile_gender_f')) { echo 'checked'; } ?>>
                                 <?= $this->lang->line('profile_gender_f'); ?>
-                            </label>
-                        </div>
-                        <?php
-                            } else {
-                        ?>
-                        <div class="form-check"> 
-                            <label for="gender" class="form-check-label">
-                                <input type="radio" class="form-check-input" id="gender_m" name="gender" value="m">
-                                <?= $this->lang->line('profile_gender_m'); ?>
                             </label>
                         </div>
                         <div class="form-check"> 
                             <label for="gender" class="form-check-label">
-                                <input type="radio" class="form-check-input" id="gender_f" name="gender" value="f" checked>
-                                <?= $this->lang->line('profile_gender_f'); ?>
+                                <input type="radio" class="form-check-input" id="no_gender" name="gender" value="" <?php if ( ! $profile_gender) { echo 'checked'; } ?>>
+                                <i><?= $this->lang->line('register_form_no_gender'); ?></i>
                             </label>
                         </div>
-                        <?php
-                            }
-                        ?>
                     </label>
                 </fieldset>
-                <label for="username" class="col-sm-4 col-form-label"><strong><?= $username; ?></strong>:
-                <input type="text" class="form-control" id="username" name="username" value="<?= $profile_username; ?>" readonly></label>
-                <label for="email" class="col-sm-4 col-form-label"><strong><?= $email; ?></strong>:
-                <input type="text" class="form-control" id="email" name="email" value="<?= $profile_email; ?>" readonly></label><br>
+                <label for="username" class="col-sm-4 col-form-label"><strong><?= $username; ?></strong>*:
+                <input type="text" class="form-control" id="username" name="username" value="<?= $profile_username; ?>"></label>
+                <label for="email" class="col-sm-4 col-form-label"><strong><?= $email; ?></strong>*:
+                <input type="text" class="form-control" id="email" name="email" value="<?= $profile_email; ?>"></label><br>
                 <label for="phone" class="col-sm-4 col-form-label"><strong><?= $phone; ?></strong>:
-                <input type="text" class="form-control" id="phone" name="phone" value="<?= $profile_phone; ?>" readonly></label>
+                <input type="text" class="form-control" id="phone" name="phone" value="<?= $profile_phone; ?>"></label>
                 <fieldset class="form-group">
-                    <label class="form-label"><strong><?= $language; ?></strong>:
-                        <?php
-                            if ($profile_language === $this->lang->line('profile_language_en')) {
-                        ?>
+                    <label class="form-label"><strong><?= $language; ?></strong>*:
                         <div class="form-check"> 
                             <label for="language" class="form-check-label">
                                 <input type="radio" class="form-check-input" id="language_en" name="language" value="en" checked>
@@ -83,24 +68,6 @@
                                 <?= $this->lang->line('profile_language_it'); ?>
                             </label>
                         </div>
-                        <?php
-                            } else {
-                        ?>
-                        <div class="form-check"> 
-                            <label for="language" class="form-check-label">
-                                <input type="radio" class="form-check-input" id="language_en" name="language" value="en">
-                                <?= $this->lang->line('profile_language_en'); ?>
-                            </label>
-                        </div>
-                        <div class="form-check"> 
-                            <label for="language" class="form-check-label">
-                                <input type="radio" class="form-check-input" id="language_it" name="language" value="it" checked>
-                                <?= $this->lang->line('profile_language_it'); ?>
-                            </label>
-                        </div>
-                        <?php
-                            }
-                        ?>
                     </label>
                 </fieldset>
             <!-- </div> -->